#12: Combine metric tests in one location
